+++ src/algorithms/wirehair_benchmark.hpp
@@ -28,7 +28,7 @@ public:
   bool check_for_corruption() const noexcept override;
 
 private:
-  uint32_t m_num_total_blocks;
+  size_t m_num_total_blocks;
 
   std::unique_ptr<uint8_t[]> m_original_buffer;    ///< Buffer for the original data we want to transmit
   std::unique_ptr<uint8_t[]> m_encode_buffer;      ///< Buffer for the encoded data
